backend/requirements.txt
# Core Django packages
Django>=5.1,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
django-filter>=24.3

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
django-oauth-toolkit>=1.8.0

# Database & Models (PostgreSQL - use psycopg2-binary for Windows)
psycopg2-binary>=2.9.9
django-extensions>=3.2.3

# File handling & Export (Updated versions)
Pillow>=10.4.0
reportlab>=4.4.4
python-docx>=1.2.0
python-pptx>=1.0.2
lxml>=6.0.2
XlsxWriter>=3.2.9
# Python-magic: platform-specific (libmagic wrapper)
python-magic>=0.4.27; sys_platform != 'win32'  # Linux/macOS
python-magic-bin>=0.4.14; sys_platform == 'win32'  # Windows
pdf2image>=1.17.0

weasyprint>=62.3  # HTML to PDF with excellent Unicode support



# Search functionality (use simpler versions or alternatives)
# django-haystack==3.2.1
# elasticsearch==7.17.7

# Caching & Background Tasks (Python 3.13 compatible)
redis>=5.1.0
django-redis>=5.4.0
celery>=5.4.0
django-celery-beat>=2.7.0

# Real-time WebSocket support
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Environment variables
python-decouple>=3.8

# Development tools
django-debug-toolbar>=4.4.0

# Testing framework
pytest>=8.3.0
pytest-django>=4.9.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0
factory-boy>=3.3.0
faker>=32.0.0

# API documentation
drf-spectacular>=0.27.0

# Email backend
# django-anymail==10.2

# File storage (AWS S3 - optional for development)
# boto3==1.29.7
# django-storages==1.14.2

# Type checking and stubs
django-stubs>=5.1.0
djangorestframework-stubs>=3.15.0
types-Pillow>=10.2.0
types-redis>=4.6.0

# Production Dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0

# Optional: For AWS S3 media storage (uncomment if using S3)
# boto3>=1.34.0
# django-storages>=1.14.2

# OCR & ML dependencies
# Note: `paddlepaddle` may require a specific wheel for GPU/CPU —
# install platform-specific wheel if needed (see https://www.paddlepaddle.org.cn/)
paddlepaddle>=2.6.0
paddleocr[doc-parser]>=2.9.0
opencv-python-headless>=4.8.0
numpy>=1.26
PyMuPDF>=1.24.0  # Faster PDF streaming, constant RAM usage
python-doctr[torch]>=0.9.0

# Vietnamese SBERT for semantic heading matching (OCR auto-fill)
sentence-transformers>=2.2.0

# ASR (Speech Recognition) - VinAI PhoWhisper
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
soundfile>=0.12.0
librosa>=0.10.0
